+++ src/PSRule/Configuration/BindingOption.cs
@@ -2,18 +2,8 @@
 // Licensed under the MIT License.
 
 using System.ComponentModel;
-using System.Diagnostics;
 
 namespace PSRule.Configuration;
-
-internal static class BindingOptionExtensions
-{
-    [DebuggerStepThrough]
-    public static StringComparer GetComparer(this BindingOption option)
-    {
-        return option.IgnoreCase.GetValueOrDefault(BindingOption.Default.IgnoreCase.Value) ? StringComparer.OrdinalIgnoreCase : StringComparer.Ordinal;
-    }
-}
 
 /// <summary>
 /// Options that affect property binding of TargetName and TargetType.
