+++ src/PSRule/Definitions/IResourceVisitor.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Definitions;
+
+internal interface IResourceVisitor
+{
+    bool Visit(IResource resource);
+}
