+++ src/PSRule/Definitions/ResourceAnnotation.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Definitions;
+
+/// <summary>
+/// A base resource annotation.
+/// </summary>
+internal abstract class ResourceAnnotation
+{
+
+}
