+++ src/PSRule/Definitions/ResourceAnnotations.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Definitions;
+
+/// <summary>
+/// Additional resource annotations.
+/// </summary>
+public sealed class ResourceAnnotations : Dictionary<string, object>
+{
+
+}
