+++ src/PSRule/Definitions/ValidateResourceAnnotation.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Definitions;
+
+/// <summary>
+/// Annotation used to flag validation issues.
+/// </summary>
+internal sealed class ValidateResourceAnnotation : ResourceAnnotation
+{
+
+}
