+++ src/PSRule/Host/HostState.cs
@@ -4,110 +4,8 @@
 using System.Management.Automation;
 using System.Management.Automation.Runspaces;
 using PSRule.Commands;
-using PSRule.Runtime;
 
 namespace PSRule.Host;
-
-/// <summary>
-/// A dynamic variable $PSRule used during Rule execution.
-/// </summary>
-internal sealed class PSRuleVariable : PSVariable
-{
-    private const string VARIABLE_NAME = "PSRule";
-
-    private readonly Runtime.PSRule _Value;
-
-    public PSRuleVariable()
-        : base(VARIABLE_NAME, null, ScopedItemOptions.ReadOnly)
-    {
-        _Value = new Runtime.PSRule(RunspaceContext.CurrentThread);
-    }
-
-    public override object Value => _Value;
-}
-
-/// <summary>
-/// A dynamic variable $Rule used during Rule execution.
-/// </summary>
-internal sealed class RuleVariable : PSVariable
-{
-    private const string VARIABLE_NAME = "Rule";
-
-    private readonly Rule _Value;
-
-    public RuleVariable()
-        : base(VARIABLE_NAME, null, ScopedItemOptions.ReadOnly)
-    {
-        _Value = new Rule();
-    }
-
-    public override object Value => _Value;
-}
-
-/// <summary>
-/// A dynamic variable $LocalizedData used during Rule execution.
-/// </summary>
-internal sealed class LocalizedDataVariable : PSVariable
-{
-    private const string VARIABLE_NAME = "LocalizedData";
-
-    private readonly LocalizedData _Value;
-
-    public LocalizedDataVariable()
-        : base(VARIABLE_NAME, null, ScopedItemOptions.ReadOnly)
-    {
-        _Value = new LocalizedData();
-    }
-
-    public override object Value => _Value;
-}
-
-/// <summary>
-/// An assertion helper variable $Assert used during Rule execution.
-/// </summary>
-internal sealed class AssertVariable : PSVariable
-{
-    private const string VARIABLE_NAME = "Assert";
-    private readonly Assert _Value;
-
-    public AssertVariable()
-        : base(VARIABLE_NAME, null, ScopedItemOptions.ReadOnly)
-    {
-        _Value = new Assert();
-    }
-
-    public override object Value => _Value;
-}
-
-/// <summary>
-/// A dynamic variable used during Rule execution.
-/// </summary>
-internal sealed class TargetObjectVariable : PSVariable
-{
-    private const string VARIABLE_NAME = "TargetObject";
-
-    public TargetObjectVariable()
-        : base(VARIABLE_NAME, null, ScopedItemOptions.ReadOnly)
-    {
-
-    }
-
-    public override object Value => RunspaceContext.CurrentThread.TargetObject?.Value;
-}
-
-internal sealed class ConfigurationVariable : PSVariable
-{
-    private const string VARIABLE_NAME = "Configuration";
-    private readonly Runtime.Configuration _Value;
-
-    public ConfigurationVariable()
-        : base(VARIABLE_NAME, null, ScopedItemOptions.ReadOnly)
-    {
-        _Value = new Runtime.Configuration(RunspaceContext.CurrentThread);
-    }
-
-    public override object Value => _Value;
-}
 
 internal static class HostState
 {
