+++ src/PSRule/Pipeline/Dependencies/LockFile.cs
@@ -3,21 +3,8 @@
 
 using System.Text;
 using Newtonsoft.Json;
-using PSRule.Data;
 
 namespace PSRule.Pipeline.Dependencies;
-
-/// <summary>
-/// An entry within the lock file.
-/// </summary>
-public sealed class LockEntry
-{
-    /// <summary>
-    /// The version to use.
-    /// </summary>
-    [JsonProperty("version", NullValueHandling = NullValueHandling.Include)]
-    public SemanticVersion.Version Version { get; set; }
-}
 
 /// <summary>
 /// Define the structure for the PSRule lock file.
