+++ src/PSRule/Pipeline/Emitters/EmitterChain.cs
@@ -3,10 +3,9 @@
 
 using PSRule.Emitters;
 
-namespace PSRule.Pipeline.Emitters
-{
-    /// <summary>
-    /// A chain of emitters.
-    /// </summary>
-    internal delegate bool EmitterChain(IEmitterContext context, object o, Type type);
-}
+namespace PSRule.Pipeline.Emitters;
+
+/// <summary>
+/// A chain of emitters.
+/// </summary>
+internal delegate bool EmitterChain(IEmitterContext context, object o, Type type);
