+++ src/PSRule/Rules/RuleHelpInfo.cs
@@ -10,108 +10,6 @@ using YamlDotNet.Serialization;
 namespace PSRule.Rules;
 
 /// <summary>
-/// A rule help information structure.
-/// </summary>
-public interface IRuleHelpInfoV2 : IResourceHelpInfo
-{
-    /// <summary>
-    /// The rule recommendation.
-    /// </summary>
-    InfoString Recommendation { get; }
-
-    /// <summary>
-    /// Additional annotations, which are string key/ value pairs.
-    /// </summary>
-    Hashtable Annotations { get; }
-
-    /// <summary>
-    /// The name of the module where the rule was loaded from.
-    /// </summary>
-    string ModuleName { get; }
-
-    /// <summary>
-    /// Additional online links to reference information for the rule.
-    /// </summary>
-    Link[] Links { get; }
-}
-
-/// <summary>
-/// Extension methods for rule help information.
-/// </summary>
-public static class RuleHelpInfoExtensions
-{
-    private const string ONLINE_HELP_LINK_ANNOTATION = "online version";
-
-    /// <summary>
-    /// Get the URI for the online version of the documentation.
-    /// </summary>
-    /// <returns>Returns the URI when a valid link is set, otherwise null is returned.</returns>
-    public static Uri GetOnlineHelpUri(this IRuleHelpInfoV2 info)
-    {
-        var link = GetOnlineHelpUrl(info);
-        return link == null ||
-            !Uri.TryCreate(link, UriKind.Absolute, out var result) ?
-            null : result;
-    }
-
-    /// <summary>
-    /// Get the URL for the online version of the documentation.
-    /// </summary>
-    /// <returns>Returns the URL when set, otherwise null is returned.</returns>
-    public static string GetOnlineHelpUrl(this IRuleHelpInfoV2 info)
-    {
-        return info == null ||
-            info.Annotations == null ||
-            !info.Annotations.ContainsKey(ONLINE_HELP_LINK_ANNOTATION) ?
-            null : info.Annotations[ONLINE_HELP_LINK_ANNOTATION].ToString();
-    }
-
-    /// <summary>
-    /// Determines if the online help link is set.
-    /// </summary>
-    /// <returns>Returns <c>true</c> when the online help link is set. Otherwise this method returns <c>false</c>.</returns>
-    internal static bool HasOnlineHelp(this IRuleHelpInfoV2 info)
-    {
-        return info != null &&
-            info.Annotations != null &&
-            info.Annotations.ContainsKey(ONLINE_HELP_LINK_ANNOTATION);
-    }
-
-    /// <summary>
-    /// Set the online help link from the <paramref name="url"/> parameter.
-    /// </summary>
-    /// <param name="info">The info object.</param>
-    /// <param name="url">A URL to the online help location.</param>
-    internal static void SetOnlineHelpUrl(this IRuleHelpInfoV2 info, string url)
-    {
-        if (info == null || info.Annotations == null || string.IsNullOrEmpty(url)) return;
-        info.Annotations[ONLINE_HELP_LINK_ANNOTATION] = url;
-    }
-}
-
-/// <summary>
-/// An URL link to reference information.
-/// </summary>
-public sealed class Link
-{
-    internal Link(string name, string uri)
-    {
-        Name = name;
-        Uri = uri;
-    }
-
-    /// <summary>
-    /// The display name of the link.
-    /// </summary>
-    public string Name { get; }
-
-    /// <summary>
-    /// The URL to the information, or the target link.
-    /// </summary>
-    public string Uri { get; }
-}
-
-/// <summary>
 /// Output view helper class for rule help.
 /// </summary>
 public sealed class RuleHelpInfo : IRuleHelpInfoV2
