+++ tests/PSRule.Tests/RuleFilterTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System;
 using System.Collections;
 using PSRule.Definitions;
 using PSRule.Definitions.Rules;
@@ -10,24 +11,24 @@ namespace PSRule;
 /// <summary>
 /// Define tests to validate <see cref="RuleFilter"/>.
 /// </summary>
-public sealed class RuleFilterTests
+public sealed partial class RuleFilterTests
 {
     [Fact]
     public void MatchInclude()
     {
-        var filter = new RuleFilter(new string[] { "rule1", "rule2" }, null, null, null, null);
-        Assert.True(filter.Match("rule1", null, null));
-        Assert.True(filter.Match("Rule2", null, null));
-        Assert.False(filter.Match("rule3", null, null));
+        var filter = new RuleFilter(["rule1", "rule2"], null, null, null, null);
+        Assert.True(filter.Match(GetResource("rule1")));
+        Assert.True(filter.Match(GetResource("Rule2")));
+        Assert.False(filter.Match(GetResource("rule3")));
     }
 
     [Fact]
     public void MatchExclude()
     {
-        var filter = new RuleFilter(null, null, new string[] { "rule3" }, null, null);
-        Assert.True(filter.Match("rule1", null, null));
-        Assert.True(filter.Match("rule2", null, null));
-        Assert.False(filter.Match("Rule3", null, null));
+        var filter = new RuleFilter(null, null, ["rule3"], null, null);
+        Assert.True(filter.Match(GetResource("rule1")));
+        Assert.True(filter.Match(GetResource("rule2")));
+        Assert.False(filter.Match(GetResource("Rule3")));
     }
 
     [Fact]
@@ -45,12 +46,23 @@ public sealed class RuleFilterTests
 
         // Check basic match
         resourceTags["category"] = "group2";
-        Assert.True(filter.Match("rule", ResourceTags.FromHashtable(resourceTags), null));
+        Assert.True(filter.Match(GetResource("rule", ResourceTags.FromHashtable(resourceTags))));
         resourceTags["category"] = "group1";
-        Assert.True(filter.Match("rule", ResourceTags.FromHashtable(resourceTags), null));
+        Assert.True(filter.Match(GetResource("rule", ResourceTags.FromHashtable(resourceTags))));
         resourceTags["category"] = "group3";
-        Assert.False(filter.Match("rule", ResourceTags.FromHashtable(resourceTags), null));
-        Assert.False(filter.Match("rule", null, null));
+        Assert.False(filter.Match(GetResource("rule", ResourceTags.FromHashtable(resourceTags))));
+        Assert.False(filter.Match(GetResource("rule")));
+
+        // Include local
+        filter = new RuleFilter(null, tag, null, true, null);
+        resourceTags["category"] = "group1";
+        Assert.True(filter.Match(GetResource("module1\\rule", ResourceTags.FromHashtable(resourceTags))));
+        resourceTags["category"] = "group3";
+        Assert.False(filter.Match(GetResource("module1\\rule", ResourceTags.FromHashtable(resourceTags))));
+        resourceTags["category"] = "group3";
+        Assert.True(filter.Match(GetResource(".\\rule", ResourceTags.FromHashtable(resourceTags))));
+        Assert.False(filter.Match(GetResource("module1\\rule")));
+        Assert.True(filter.Match(GetResource(".\\rule")));
     }
 
     [Fact]
@@ -62,19 +74,28 @@ public sealed class RuleFilterTests
         // Create a filter
         var labels = new ResourceLabels
         {
-            ["framework.v1/control"] = new string[] { "c-1", "c-2" }
+            ["framework.v1/control"] = ["c-1", "c-2"]
         };
         var filter = new RuleFilter(null, null, null, null, labels);
 
         resourceLabels["framework.v1/control"] = new string[] { "c-2", "c-1" };
-        Assert.True(filter.Match("rule", null, ResourceLabels.FromHashtable(resourceLabels)));
+        Assert.True(filter.Match(GetResource("rule", null, ResourceLabels.FromHashtable(resourceLabels))));
         resourceLabels["framework.v1/control"] = new string[] { "c-3", "c-1" };
-        Assert.True(filter.Match("rule", null, ResourceLabels.FromHashtable(resourceLabels)));
+        Assert.True(filter.Match(GetResource("rule", null, ResourceLabels.FromHashtable(resourceLabels))));
         resourceLabels["framework.v1/control"] = new string[] { "c-1", "c-3" };
-        Assert.True(filter.Match("rule", null, ResourceLabels.FromHashtable(resourceLabels)));
+        Assert.True(filter.Match(GetResource("rule", null, ResourceLabels.FromHashtable(resourceLabels))));
         resourceLabels["framework.v1/control"] = new string[] { "c-3", "c-4" };
-        Assert.False(filter.Match("rule", null, ResourceLabels.FromHashtable(resourceLabels)));
-        resourceLabels["framework.v1/control"] = System.Array.Empty<string>();
-        Assert.False(filter.Match("rule", null, ResourceLabels.FromHashtable(resourceLabels)));
+        Assert.False(filter.Match(GetResource("rule", null, ResourceLabels.FromHashtable(resourceLabels))));
+        resourceLabels["framework.v1/control"] = Array.Empty<string>();
+        Assert.False(filter.Match(GetResource("rule", null, ResourceLabels.FromHashtable(resourceLabels))));
     }
+
+    #region Helper Methods
+
+    private static IResource GetResource(string id, ResourceTags resourceTags = null, ResourceLabels resourceLabels = null)
+    {
+        return new TestResourceName(ResourceId.Parse(id), resourceTags, resourceLabels);
+    }
+
+    #endregion Helper Methods
 }
