+++ tests/PSRule.Tests/Usings.cs
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+global using Moq;
 global using Xunit;
 global using Assert = Xunit.Assert;
-global using Moq;
